+++ couch_rs/src/types/index.rs
@@ -1,19 +1,17 @@
 use super::*;
-use serde::{Serialize, Deserialize};
-use find::{SortSpec};
-use document::{DocumentId};
+use document::DocumentId;
+use find::SortSpec;
+use serde::{Deserialize, Serialize};
 
 /// Index fields abstraction
 #[derive(Serialize, Deserialize, Eq, PartialEq, Debug, Clone)]
 pub struct IndexFields {
-    pub fields: Vec<SortSpec>
+    pub fields: Vec<SortSpec>,
 }
 
 impl IndexFields {
     pub fn new(fields: Vec<SortSpec>) -> IndexFields {
-        IndexFields {
-            fields
-        }
+        IndexFields { fields }
     }
 }
 
@@ -24,12 +22,12 @@ pub struct Index {
     pub name: String,
     #[serde(rename = "type")]
     pub index_type: String,
-    pub def: IndexFields
+    pub def: IndexFields,
 }
 
 /// Database index list abstraction
 #[derive(Serialize, Deserialize, PartialEq, Debug, Clone)]
 pub struct DatabaseIndexList {
     pub total_rows: u32,
-    pub indexes: Vec<Index>
+    pub indexes: Vec<Index>,
 }